Restore simplified markerarray for openmower-gui
The true path can be very large, 90k+ points for one of the areas on my map.
This appears to exceed a limit in the linestring renderer used in openmower-gui.

Restore the old inaccurate/smaller makerarray for openmower-gui and add
a new accurate markerarray topic for use in rviz (disabled by default, can be
enabled for debugging/sim).