Fix use-after-free
`areaLastPoint` used to be a pointer to the last element of the vector
returned by `line.equally_spaced_points(...)`. However, that vector
is only scoped within the for-loop, which means that at the time when
the pointer is used to detect the closest obstacle, the memory might
have been overwritten. This actually happened to me in about one out of
three cases, leading to bogus start coordinates and therefore unstable
obstacle ordering.

Using the existence of at least one area outline should be sufficient
to signal that the point can be used and is no longer the initial (0,0).